+++ src/PaymentManager.php
@@ -8,6 +8,7 @@ use AliYavari\IranPayment\Contracts\Payment;
 use AliYavari\IranPayment\Drivers\BehpardakhtDriver;
 use AliYavari\IranPayment\Drivers\FakeDriver;
 use AliYavari\IranPayment\Drivers\IdpayDriver;
+use AliYavari\IranPayment\Drivers\NextpayDriver;
 use AliYavari\IranPayment\Drivers\PaypingDriver;
 use AliYavari\IranPayment\Drivers\PepDriver;
 use AliYavari\IranPayment\Drivers\SadadDriver;
@@ -92,6 +93,11 @@ final class PaymentManager extends Manager
     protected function createPaypingDriver(): PaypingDriver
     {
         return $this->container->make(PaypingDriver::class);
+    }
+
+    protected function createNextpayDriver(): NextpayDriver
+    {
+        return $this->container->make(NextpayDriver::class);
     }
 
     private function shouldBeImmutable(string $driver): bool