+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - **Zibal** gateway
 - **Payping** gateway
+- **Nextpay** gateway
 
 ### Changed